Fixed issue with partials not loading
Resolves an issue where partials are not loaded correctly in the event
the `options.cache` is set to `true`. Accordingly allows partials to be
loaded once, or loaded every time if caching is disabled.

In order to test both situations where caching may be on or off, we
toggle the `process.env.NODE_ENV` variable which in effect toggles the
`options.cache` parameter. Thus I have included grunt to run the same
test suite twice; once as the test suite was run before, and then again
with the environment variable set to production.