In the Example, illustrate use of "application local variable" to configure "if" helper.
I have a Python/Zope/Plone background, and hence I'm not accustomed to the "convention over configuration" paradigm that rules in Node.js. Such as the (partly undocumented) voodoo with "views/index.hbs" and "{{{body}}}". Or how the Express "app.locals" influences "settings" in a view.

Hopefully, my little contribution to the Example will save a few minutes for some fellow rookies.