Pass correct context to i18n helper
The i18n library requires the "this" keyword to point to the root
context, which has the current locale attached to it.  Inside of
an each block or other context where the "this" keyword points to
something else, translations will fail because it cannot find the
correct locale, unless we explicitly reference the root context.