Added support for a restrictLayoutsTo option
refs https://github.com/TryGhost/Product/issues/4191

Passing this option will cause express-hbs to error when attempting to read a
layout from outside of the path. This is used to help alleviate the security
concerns when using user submitted data for choosing layouts

Long term, in a major release we can use the `layoutsDir` option for this, but
we did not want to make a breaking change at this stage.